api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pyyaml==6.0.2
PyJWT==2.10.1
httpx==0.28.1
python-multipart==0.0.20
